+++ lib/stringx.mli
@@ -538,3 +538,37 @@ val to_snake_case : string -> string
     - to_snake_case "HTTP20xOK" = "http_20x_ok"
     - to_snake_case "Duration2m3s" = "duration_2m3s"
     - to_snake_case "Bld4Floor3rd" = "bld4_floor_3rd" *)
+
+val map : (Uchar.t -> Uchar.t) -> string -> string
+(** [map f s] returns a new string which is the result of applying [f] to each
+    Unicode code point of [s]. The mapping function [f] must return a valid
+    [Uchar.t] for every input.
+
+    This function is Unicode-aware: it decodes [s] into code points, applies
+    [f], then re-encodes into UTF-8.
+
+    Example: let rot13 u = let c = Uchar.to_int u in if c >= Char.code 'A' && c
+    <= Char.code 'Z' then Uchar.of_int (Char.code 'A' + ((c - Char.code 'A' +
+    13) mod 26)) else if c >= Char.code 'a' && c <= Char.code 'z' then
+    Uchar.of_int (Char.code 'a' + ((c - Char.code 'a' + 13) mod 26)) else u in
+    map rot13 "'Twas brillig and the slithy camel..." = "'Gjnf oevyyvt naq gur
+    fyvgul pnzry..." *)
+
+val filter_map : (Uchar.t -> Uchar.t option) -> string -> string
+(** [filter_map f s] applies [f] to each Unicode code point [u] of [s]. If [f u]
+    returns [Some u'], [u'] is kept in the result; if [None], [u] is dropped.
+
+    This function is Unicode-aware: it decodes [s] into code points, applies
+    [f], then re-encodes into UTF-8.
+
+    Example: let drop_vowel u = match Uchar.to_int u with | c when List.mem c
+    [ Char.code 'a'; Char.code 'e'; Char.code 'i' ; Char.code 'o'; Char.code 'u'
+     ] -> None | _ -> Some u in filter_map drop_vowel "hello" = "hll" *)
+
+val iter : (Uchar.t -> unit) -> string -> unit
+(** [iter f s] applies [f] to each Unicode code point of [s], in sequence,
+    purely for side-effects. *)
+
+val fold : ('acc -> Uchar.t -> 'acc) -> 'acc -> string -> 'acc
+(** [fold f init s] applies [f acc u] to each Unicode code point [u] of [s],
+    carrying along an accumulator [acc], and returns the final accumulator. *)